requirements.txt
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.8.0
cryptography==2.3.1
cycler==0.10.0
Cython==0.29.6
dask==1.1.4
decorator==4.4.0
defusedxml==0.5.0
entrypoints==0.3
graphviz==0.8.4
h5py==2.8.0
h5py-cache==1.0
idna==2.7
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
Jinja2==2.10
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kiwisolver==1.0.1
lmdb~=1.5.1
MarkupSafe==1.1.1
matplotlib~=3.9.2
mistune==0.8.4
mkl-fft==1.0.6
mkl-random==1.0.1
nbconvert==5.4.1
nbformat==4.4.0
networkx~=3.2.1
notebook==5.7.6
numpy~=1.26.4
olefile==0.46
pandas~=2.2.3
pandocfilters==1.4.2
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
Pillow~=9.4.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
pyOpenSSL==18.0.0
pyparsing==2.3.1
pyrsistent==0.14.11
PySocks==1.6.8
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
pyzmq==17.1.2
requests==2.20.1
scikit-image~=0.24.0
scipy~=1.13.1
Send2Trash==1.5.0
Shapely~=2.0.6
six==1.11.0
svgpathtools~=1.6.1
svgwrite==1.2.1
tensorboardX~=2.6.2.2
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==1.0.0
torchfile==0.1.0
torchnet==0.0.4
torchvision~=0.19.1
tornado==5.1.1
tqdm~=4.66.5
traitlets==4.3.2
urllib3==1.24.1
visdom==0.1.8.5
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0

opencv-python~=4.10.0.84
triangle~=20250106
cairosvg~=2.7.1